requirements.txt
python-telegram-bot==20.0
firebase-admin==6.2.0
python-decouple==3.8  